backend/requirements.txt
# ============================================
# Recruitr Backend Dependencies
# ============================================

# Web Framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6

# Database & ORM
supabase==2.0.3
postgrest==0.13.2

# Information Retrieval
sentence-transformers==2.2.2
rank-bm25==0.2.2
nltk==3.8.1

# Machine Learning & Data
numpy==1.24.3
scikit-learn==1.3.2

# LLM APIs
google-generativeai==0.3.2

# Environment & Configuration
python-dotenv==1.0.0
pydantic==2.5.0
pydantic-settings==2.1.0

# HTTP & API
httpx==0.24.1
requests==2.31.0

# Utilities
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4

